+++ src/wai/annotations/video/isp/filter_frames_by_label/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package for the filter-frames-by-label-od ISP.
+"""