+++ src/wai/annotations/video/isp/filter_frames_by_label/component/__init__.py
@@ -1,0 +1,1 @@
+from ._FilterFramesByLabel import FilterFramesByLabel